+++ src/theme/BlogPostItem/index.js
@@ -1,16 +1,22 @@
 import React from 'react';
-import BlogPostItem from '@theme-original/BlogPostItem';
-import Comments from '../../components/Comments'
+import clsx from 'clsx';
 import {useBlogPost} from '@docusaurus/theme-common/internal';
-
-
-export default function BlogPostItemWrapper(props) { 
-  const {isBlogPostPage} = useBlogPost()
- 
+import BlogPostItemContainer from '@theme/BlogPostItem/Container';
+import BlogPostItemHeader from '@theme/BlogPostItem/Header';
+import BlogPostItemContent from '@theme/BlogPostItem/Content';
+import BlogPostItemFooter from '@theme/BlogPostItem/Footer';
+// apply a bottom margin in list view
+function useContainerClassName() {
+  const {isBlogPostPage} = useBlogPost();
+  return !isBlogPostPage ? 'margin-bottom--xl' : undefined;
+}
+export default function BlogPostItem({children, className}) {
+  const containerClassName = useContainerClassName();
   return (
-    <>
-      <BlogPostItem {...props} />
-      {isBlogPostPage && <Comments {...{type:'blog'}}/>}
-    </>
+    <BlogPostItemContainer className={clsx(containerClassName, className)}>
+      <BlogPostItemHeader />
+      <BlogPostItemContent>{children}</BlogPostItemContent>
+      <BlogPostItemFooter />
+    </BlogPostItemContainer>
   );
 }